Added support for modern MSVS solution/projects
Specifically adding support for sln solution files and the proj/xproj
files used since MS Visual Studio 8 and above. This is the basic set of
features, and from here on is to refine the capabilities and output to
better align with the most probable intentions of those whom created the
solutions in the first place.

First targets will be support for vfproj for Fortran, and vdproj for
setup projects.